requirements.txt
parsedatetime==2.5
Pillow==7.0.0
